fix: do not export anymore from the App.tsx file!
