+++ source/sites/publicodes/App.tsx
@@ -33,6 +33,7 @@ import Diffuser from './pages/Diffuser'
 import Profil from './Profil.tsx'
 import sitePaths from './sitePaths'
 import TranslationContribution from './TranslationContribution'
+import { isFluidLayout } from './utils'
 
 // All those lazy components, could be probably be handled another more consise way
 // Also, see this issue about migrating to SSR https://github.com/datagir/nosgestesclimat-site/issues/801
@@ -106,16 +107,6 @@ export default function Root({}) {
 		>
 			<Main />
 		</Provider>
-	)
-}
-
-export const isFluidLayout = (encodedPathname) => {
-	const pathname = decodeURIComponent(encodedPathname)
-
-	return (
-		pathname === '/' ||
-		pathname.startsWith('/nouveautés') ||
-		pathname.startsWith('/documentation')
 	)
 }
 
@@ -216,8 +207,6 @@ const Main = ({}) => {
 	)
 }
 
-export const Loading = AnimatedLoader
-
 const Router = ({}) => {
 	return (
 		<Routes>
@@ -226,7 +215,7 @@ const Router = ({}) => {
 				path="documentation/*"
 				element={
 					<WithEngine options={{ parsed: false, optimized: false }}>
-						<Suspense fallback={<Loading />}>
+						<Suspense fallback={<AnimatedLoader />}>
 							<DocumentationLazy />
 						</Suspense>
 					</WithEngine>
@@ -235,7 +224,7 @@ const Router = ({}) => {
 			<Route
 				path={encodeURIComponent('modèle')}
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<ModelLazy />
 					</Suspense>
 				}
@@ -243,7 +232,7 @@ const Router = ({}) => {
 			<Route
 				path="simulateur/*"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<WithEngine>
 							<SimulateurLazy />
 						</WithEngine>
@@ -253,7 +242,7 @@ const Router = ({}) => {
 			<Route
 				path="/stats"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<StatsLazy />
 					</Suspense>
 				}
@@ -261,7 +250,7 @@ const Router = ({}) => {
 			<Route
 				path="/fin/*"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<WithEngine>
 							<FinLazy />
 						</WithEngine>
@@ -271,7 +260,7 @@ const Router = ({}) => {
 			<Route
 				path="/personas"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<WithEngine>
 							<PersonasLazy />
 						</WithEngine>
@@ -281,7 +270,7 @@ const Router = ({}) => {
 			<Route
 				path="/actions/*"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<WithEngine>
 							<ActionsLazy />
 						</WithEngine>
@@ -299,7 +288,7 @@ const Router = ({}) => {
 			<Route
 				path="/contribuer/*"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<ContributionLazy />
 					</Suspense>
 				}
@@ -307,7 +296,7 @@ const Router = ({}) => {
 			<Route
 				path="/contribuer-traduction"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<TranslationContribution />
 					</Suspense>
 				}
@@ -346,7 +335,7 @@ const Router = ({}) => {
 			<Route
 				path={`${encodeURIComponent('nouveautés')}/*`}
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<News />
 					</Suspense>
 				}
@@ -354,7 +343,7 @@ const Router = ({}) => {
 			<Route
 				path="/guide"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<GuideGroupeLazy />
 					</Suspense>
 				}
@@ -362,7 +351,7 @@ const Router = ({}) => {
 			<Route
 				path="/guide/:encodedName"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<GuideGroupeLazy />
 					</Suspense>
 				}
@@ -370,7 +359,7 @@ const Router = ({}) => {
 			<Route
 				path={`${encodeURIComponent('conférence')}/:room`}
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<ConferenceLazy />
 					</Suspense>
 				}
@@ -378,7 +367,7 @@ const Router = ({}) => {
 			<Route
 				path="/groupe"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<GroupSwitchLazy />
 					</Suspense>
 				}
@@ -386,7 +375,7 @@ const Router = ({}) => {
 			<Route
 				path="/groupe/documentation-contexte"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<DocumentationContexteLazy />
 					</Suspense>
 				}
@@ -394,7 +383,7 @@ const Router = ({}) => {
 			<Route
 				path="/sondage/:room"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<SurveyLazy />
 					</Suspense>
 				}
@@ -402,7 +391,7 @@ const Router = ({}) => {
 			<Route
 				path="/accessibilite"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<AccessibilityLazy />
 					</Suspense>
 				}
@@ -410,7 +399,7 @@ const Router = ({}) => {
 			<Route
 				path="/tutoriel"
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<TutorialLazy />
 					</Suspense>
 				}
@@ -418,7 +407,7 @@ const Router = ({}) => {
 			<Route
 				path={`/${encodeURIComponent('pétrole-et-gaz')}`}
 				element={
-					<Suspense fallback={<Loading />}>
+					<Suspense fallback={<AnimatedLoader />}>
 						<PetrogazLandingLazy />
 					</Suspense>
 				}
