+++ source/sites/publicodes/utils.tsx
@@ -1,0 +1,9 @@
+export const isFluidLayout = (encodedPathname) => {
+	const pathname = decodeURIComponent(encodedPathname)
+
+	return (
+		pathname === '/' ||
+		pathname.startsWith('/nouveautés') ||
+		pathname.startsWith('/documentation')
+	)
+}
